+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/meal/ui/card/MealsCard.kt
@@ -170,58 +170,8 @@ private fun HorizontalMealsCard(
     }
 }
 
-@OptIn(ExperimentalSharedTransitionApi::class, ExperimentalAnimationApi::class)
 @Composable
-private fun VerticalMealsCard(
-    meals: List<MealWithSummary>?,
-    onMealClick: (mealId: Long) -> Unit,
-    onAddClick: (mealId: Long) -> Unit,
-    onLongClick: () -> Unit,
-    animatedVisibilityScope: AnimatedVisibilityScope,
-    epochDay: Int,
-    contentPadding: PaddingValues,
-    shimmer: Shimmer,
-    modifier: Modifier = Modifier
-) = with(LocalNavigationSharedTransitionScope.current ?: error("SharedTransitionScope not found")) {
-    val transition = updateTransition(meals)
-
-    val meals = remember(meals) {
-        meals ?: List<MealWithSummary?>(4) { null }
-    }
-
-    Column(
-        modifier = modifier.padding(contentPadding),
-        verticalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        meals.forEachIndexed { i, meal ->
-            transition.Crossfade(
-                contentKey = { it != null },
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                if (it != null && meal != null) {
-                    MealCard(
-                        animatedVisibilityScope = animatedVisibilityScope,
-                        epochDay = epochDay,
-                        meal = meal,
-                        isEmpty = meal.isEmpty,
-                        totalCalories = meal.calories,
-                        totalProteins = meal.proteins,
-                        totalCarbohydrates = meal.carbohydrates,
-                        totalFats = meal.fats,
-                        onMealClick = { onMealClick(meal.id) },
-                        onAddClick = { onAddClick(meal.id) },
-                        onLongClick = onLongClick
-                    )
-                } else {
-                    MealCardSkeleton(shimmer)
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun MealCardSkeleton(shimmer: Shimmer, modifier: Modifier = Modifier) {
+internal fun MealCardSkeleton(shimmer: Shimmer, modifier: Modifier = Modifier) {
     val headline = @Composable {
         Column {
             Box(
@@ -284,7 +234,7 @@ private fun MealCardSkeleton(shimmer: Shimmer, modifier: Modifier = Modifier) {
 
 @OptIn(ExperimentalSharedTransitionApi::class)
 @Composable
-private fun SharedTransitionScope.MealCard(
+internal fun SharedTransitionScope.MealCard(
     animatedVisibilityScope: AnimatedVisibilityScope,
     epochDay: Int,
     meal: MealWithSummary,
